requirements.txt
# Danh sách thư viện
pandas
PyYAML
openpyxl